pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bmad-assist"
version = "0.4.25"
description = "CLI tool for automating BMAD methodology development loop"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Pawel", email = "pawel@wizjonarium.pl"}
]
dependencies = [
    "typer>=0.21.1",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18.0,<1.0.0",  # REQUIRED: comment preservation for sprint-status.yaml
    "python-frontmatter>=1.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "claude-agent-sdk>=0.1.0",
    "textstat>=0.7.3",
    "httpx>=0.27.0",
    "questionary>=2.0.0",  # Interactive prompts with arrow-key navigation
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-PyYAML>=6.0.0",
    "scipy-stubs>=1.14.0",
]

[project.scripts]
bmad-assist = "bmad_assist.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
bmad_assist = ["workflows/**/*", "workflows/cache/*.tpl.xml", "workflows/cache/*.meta.yaml", "default_patches/*.yaml", "testarch/knowledge_base/**/*", "deep_verify/knowledge/data/*.yaml"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

# External libraries without py.typed marker or stubs
# These are legitimate third-party libs that don't ship type info
[[tool.mypy.overrides]]
module = [
    "frontmatter",
    "textstat",
    "playwright.*",
    "watchdog.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D", "UP", "B", "C4", "SIM"]
ignore = [
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D203",   # 1 blank line required before class docstring
    "D213",   # Multi-line docstring summary should start at the second line
    "D401",   # First line of docstring should be in imperative mood
    "D417",   # Missing argument descriptions in docstring
    "B008",   # Do not perform function call in argument defaults (typer pattern)
    "B007",   # Loop control variable not used within loop body
    "B009",   # Do not call getattr with a constant attribute value
    "E501",   # Line too long - handled by formatter
    "E402",   # Module level import not at top of file
    "SIM108", # Use ternary operator instead of if-else-block
    "SIM116", # Use a dictionary instead of consecutive if statements
    "SIM103", # Return the condition directly
    "UP042", # Prefer StrEnum over (str, Enum) - keep existing pattern for consistency
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
# Default: skip slow tests for fast feedback (run slow with: pytest -m slow)
addopts = "-v --tb=short --ignore=tests/fixtures -m \"not slow\""
# LLM-optimized (token-saving): pytest -q --tb=line --no-header
markers = [
    "real_debug_logger: mark test to use real DebugJsonLogger instead of mock",
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (>5s) - deselect with '-m \"not slow\"'",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src/bmad_assist"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    # Dead code - old modules replaced by subdirectory versions
    # TODO: Delete these in separate branch (see docs/test-refactor/TEST-HEALTH.md)
    "src/bmad_assist/output.py",
    "src/bmad_assist/generator.py",
    "src/bmad_assist/executor.py",
    "src/bmad_assist/checker.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
show_missing = true
skip_covered = false
fail_under = 0

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.0",
]
